refactored App.js to be in PostsPage.js and reimplemented spread operator in addCommetn as the other syntax was no longer clearing the input value after submission
