+++ instagram/src/App.js
@@ -1,46 +1,23 @@
 import React, { Component } from 'react';
 import './App.css';
-import dummyData from './dummy-data';
-import SearchBarContainer from './components/SearchBar/SearchBarContainer';
-import PostContainer from './components/PostContainer/PostContainer';
+import PostsPage from './components/PostContainer/PostsPage';
+
+
 
 class App extends Component {
-  constructor() {
-    super();
-    this.state = {
-      posts: [],
-      filteredPosts: []
-    };
-  }
-
-  componentDidMount() {
-    let storedPosts = JSON.parse(localStorage.getItem('posts'));
-    if(storedPosts) {this.setState({ posts: storedPosts})}
-    this.setState({
-      posts: dummyData,
-    })
-  }
-
-  handleSearchChange = e => {
-    let posts = this.state.posts.filter(post => {
-      if(post.username.includes(e.target.value)) {
-        return post;
-      }
-    });return this.setState({ filteredPosts: posts})
-  }
+  // constructor() {
+  //   super();
+  //   this.state = {
+  //     posts: [],
+  //     filteredPosts: []
+  //   };
+  // }
 
   render() {
     
     return (
       <div className="App">
-        <SearchBarContainer 
-          searchPosts={this.handleSearchChange}
-        />
-        <PostContainer 
-          data={this.state.filteredPosts.length > 0 ?
-            this.state.filteredPosts : 
-            this.state.posts}
-        />
+        <PostsPage />
       </div>
     );
   }
