+++ instagram/src/components/CommentSection/CommentCard.js
@@ -6,7 +6,6 @@ import CommentInput from './CommentInput';
 
 let moment = require('moment');
 
-// {console.log("COMMENTCARD PROPS", props)}
 class CommentCard extends Component {
   constructor(props) {
     super(props);
@@ -17,51 +16,37 @@ class CommentCard extends Component {
     }
   }
   
-  componentDidMount() {
-    console.log(this.state.comments);
+  // componentDidMount() {
+  //   console.log(this.state.comments);
     
     // let storedComments = JSON.parse(localStorage.getItem('comments'));
     // if(storedComments) {this.setState({ comments: storedComments})}
     // if(storedComments) {this.setState({ comments: this.state.comments})}
-    }
+    // }
     
     
     addComment = e => {
       e.preventDefault();
       // console.log(...this.state.comments);
       
-      // let newComments = [...this.state.comments,
-      //   {username: "Ghosty", text: this.state.comment}];
-      //   localStorage.setItem('comments', JSON.stringify(newComments))
+      let newComments = [...this.state.comments,
+        {username: "Ghosty", text: this.state.comment}];
+        localStorage.setItem('comments', JSON.stringify(newComments))
       
-      // this.setState({
-      //   comments: [...this.state.comments,
-      //     {
-      //       username: "Ghosty",
-      //       text: this.state.comment,
-
-      //     }],
-      //     comment: '',
-      //   })
-
-      let commentsCopy = this.state.comments.slice();
-      commentsCopy.push({
-        username: "Ghosty",
-        text: this.state.comment,
-      })
-      // console.log(commentsCopy);
       this.setState({
-        comments: commentsCopy 
-      })
-
+        comments: [...this.state.comments,
+          {
+            username: "Ghosty",
+            text: this.state.comment,
+          }],
+          comment: '',
+        })
       }
       
       handleChange = e => {
-        // console.log("changing to: ", e.target.value);
-        // console.log("changing input name: ", e.target.name);
         this.setState({
           [e.target.name]: e.target.value,
-        })    
+        })
       }
       
       render () {
