+++ instagram/src/components/PostContainer/PostCard.js
@@ -9,27 +9,7 @@ class PostCard extends Component {
     };
   }
 
-  // handleToggleLike = (text) => {
-    
-  //   // let likes = this.state.likes;
-    
-  //   this.setState(prevState => (
-  //     {
-  //       likes: prevState.likes.map(post => {
-  //         if (text === post.text) {
-  //           return {
-  //             ...post,
-  //             likes: post.likes + 1
-  //           }
-  //         }
-  //       })
-  //     }
-  //   ))
-  // }
-
-  
   toggleLike = (e) => {
-    // console.log("EVENT", e);
     if(e.target.classList.contains('liked')){
       e.target.classList.toggle('liked');
       let likes = this.state.likes - 1;
@@ -39,7 +19,19 @@ class PostCard extends Component {
     let likes = this.state.likes + 1;
     this.setState({ likes })}
   };
-  
+
+  toggleComment = (e) => {
+  //   if(e.target.classList.contains('no-comment')){
+  //     e.target.classList.toggle('no-comment');
+  //     let likes = this.state.likes - 1;
+  //     this.setState({ likes });
+  //   }
+  //   else {e.target.classList.toggle('liked');
+  //   let likes = this.state.likes + 1;
+  //   this.setState({ likes })
+  // }
+  };
+
   render () {
     console.log("POSTCARD PROPS", this.props);
     return (
@@ -65,7 +57,9 @@ class PostCard extends Component {
             className="far fa-heart"
             onClick={this.toggleLike}
           ></i>
-          <i className="far fa-comment"></i>
+          <i className="far fa-comment"
+            onClick={this.toggleComment}
+          ></i>
         </div>
         <p>{this.state.likes} likes</p>
         <CommentCard
