+++ instagram/src/components/PostContainer/PostContainer.css
@@ -59,3 +59,7 @@
 .liked {
   color: red;
 }
+
+.no-comment {
+  display:none
+}
